+++ apps/io-services-cms-webapp/src/utils/ingestion/__tests__/ingestion-handlers.test.ts
@@ -7,6 +7,7 @@ import { pipe } from "fp-ts/lib/function";
 import { Json } from "io-ts-types";
 import { afterEach, describe, expect, it, vi } from "vitest";
 import {
+  createIngestionBlobTriggerHandler,
   createIngestionCosmosDBTriggerHandler,
   createIngestionRetryQueueTriggerHandler,
 } from "../ingestion-handlers";
@@ -36,8 +37,9 @@ const aFormatter = vi.fn((item: ItemType) =>
   }),
 );
 
-const aFormatterWhichFails = (item: ItemType) =>
-  E.left(new Error("Failed to format item"));
+const aFormatterWhichFails = vi.fn((item: ItemType) =>
+  E.left(new Error("Failed to format item")),
+);
 
 const aProducer = {
   sendBatch: vi.fn(() => Promise.resolve()),
@@ -101,8 +103,8 @@ describe("Generic Ingestion PDND Handlers", () => {
         aFormatter,
       )(context, anInvalidQueueItem);
 
-      expect(aFormatter).toBeCalledTimes(0);
-      expect(aProducer.sendBatch).toBeCalledTimes(0);
+      expect(aFormatter).not.toBeCalled();
+      expect(aProducer.sendBatch).not.toBeCalled();
     });
 
     it("should not reject on retry succes", async () => {
@@ -114,8 +116,10 @@ describe("Generic Ingestion PDND Handlers", () => {
         aFormatter,
       )(context, aValidQueueItem);
 
-      expect(aFormatter).toBeCalledTimes(1);
-      expect(aProducer.sendBatch).toBeCalledTimes(1);
+      expect(aFormatter).toHaveBeenCalledOnce();
+      expect(aFormatter).toBeCalledWith(anItem);
+      expect(aProducer.sendBatch).toHaveBeenCalledOnce();
+      expect(aProducer.sendBatch).toBeCalledWith([{ body: anItem }]);
     });
 
     it("should throw on non permanent error(transient error)", async () => {
@@ -128,6 +132,61 @@ describe("Generic Ingestion PDND Handlers", () => {
           aFormatter,
         )(context, aValidQueueItem),
       ).rejects.toThrowError("Failed to send batch");
+    });
+  });
+
+  describe("Blob Trigger", async () => {
+    it("should success when all elements are sent to eventhub", async () => {
+      //given when
+      const res = await createIngestionBlobTriggerHandler(
+        aProducer,
+        aFormatter,
+      )({ items: [anItem] })();
+
+      //then
+      expect(E.isRight(res)).toBeTruthy();
+      if (E.isRight(res)) {
+        expect(res.right).toStrictEqual([{}]);
+      }
+      expect(aFormatter).toHaveBeenCalledOnce();
+      expect(aFormatter).toBeCalledWith(anItem);
+      expect(aProducer.sendBatch).toHaveBeenCalledOnce();
+      expect(aProducer.sendBatch).toBeCalledWith([{ body: anItem }]);
+    });
+
+    it("should throw on eventhub error", async () => {
+      //given when
+      const res = await createIngestionBlobTriggerHandler(
+        aProducerWhichFails,
+        aFormatter,
+      )({ items: [anItem] })();
+
+      //then
+      expect(E.isLeft(res)).toBeTruthy();
+      if (E.isLeft(res)) {
+        expect(res.left.message).toBe("Failed to send batch");
+      }
+      expect(aFormatter).toHaveBeenCalledOnce();
+      expect(aFormatter).toBeCalledWith(anItem);
+      expect(aProducerWhichFails.sendBatch).toHaveBeenCalledOnce();
+      expect(aProducerWhichFails.sendBatch).toBeCalledWith([{ body: anItem }]);
+    });
+
+    it("should throw on avro formatter error", async () => {
+      //given when
+      const res = await createIngestionBlobTriggerHandler(
+        aProducer,
+        aFormatterWhichFails,
+      )({ items: [anItem] })();
+
+      //then
+      expect(E.isLeft(res)).toBeTruthy();
+      if (E.isLeft(res)) {
+        expect(res.left.message).toBe("Failed to format item");
+      }
+      expect(aFormatterWhichFails).toHaveBeenCalledOnce();
+      expect(aFormatterWhichFails).toBeCalledWith(anItem);
+      expect(aProducerWhichFails.sendBatch).not.toBeCalled();
     });
   });
 });